Move guacamole code to sessions package
