+++ backend/t4cclient/sessions/__init__.py
@@ -1,3 +1,4 @@
 # Copyright DB Netz AG and the capella-collab-manager contributors
 # SPDX-License-Identifier: Apache-2.0
 
+import t4cclient.sessions.models
