+++ backend/t4cclient/sql_models/__init__.py
@@ -2,5 +2,4 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # These import statements of the models are required and should not be removed! (SQLAlchemy will not load the models otherwise)
-from ..sessions import models
 from . import notices, repositories, users
